Test Flash API Ok

+++ Common/Platform_Types.h
@@ -85,15 +85,15 @@ typedef signed char           sint8;         /*        -128 .. +127           */
 typedef unsigned char         uint8;         /*           0 .. 255            */
 typedef signed short          sint16;        /*      -32768 .. +32767         */
 typedef unsigned short        uint16;        /*           0 .. 65535          */
-typedef signed long           sint32;        /* -2147483648 .. +2147483647    */
-typedef unsigned long         uint32;        /*           0 .. 4294967295     */
+typedef signed int           sint32;        /* -2147483648 .. +2147483647    */
+typedef unsigned int          uint32;        /*           0 .. 4294967295     */
 
 typedef signed char           sint8_least;   /* At least 7 bit + 1 bit sign   */
 typedef unsigned char         uint8_least;   /* At least 8 bit                */
 typedef signed short          sint16_least;  /* At least 15 bit + 1 bit sign  */
 typedef unsigned short        uint16_least;  /* At least 16 bit               */
-typedef signed long           sint32_least;  /* At least 31 bit + 1 bit sign  */
-typedef unsigned long         uint32_least;  /* At least 32 bit               */
+typedef signed int           sint32_least;  /* At least 31 bit + 1 bit sign  */
+typedef unsigned int          uint32_least;  /* At least 32 bit               */
 
 typedef float                 float32;
 typedef double                float64;
